refactor: Improve array operations and update type imports
- useActionRecommendationOverlay.ts: forEach+push → map+spread, for...of → forEach
- Import ValidatedActionPrices from @/domain/types
- StockChart.tsx: Import ValidatedActionPrices from @/domain/types
